-Create BaseActivity, which has the toolbar and all common function -R
