+++ app/src/main/java/com/example/amcc/view/BaseActivity.java
@@ -1,0 +1,92 @@
+package com.example.amcc.view;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.view.Menu;
+import android.view.MenuItem;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
+import com.example.amcc.R;
+
+public class BaseActivity extends AppCompatActivity {
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.app_menu, menu);
+        return true;
+    }
+
+    private void shareIt() {
+        //sharing implementation here
+        Intent sharingIntent = new Intent(Intent.ACTION_SEND);
+        sharingIntent.setType("text/plain");
+        String shareBody = "Thank you for sharing our App.";
+        sharingIntent.putExtra(Intent.EXTRA_SUBJECT, "Share it");
+        sharingIntent.putExtra(android.content.Intent.EXTRA_TEXT, shareBody);
+        startActivity(Intent.createChooser(sharingIntent, null));
+    }
+
+    private void setupSetting() {
+        // set up Settings activity
+        Intent settingsIntent = new Intent(this, SettingsActivity.class);
+        startActivity(settingsIntent);
+    }
+
+    public void createToolBar() {
+
+        // Find the toolbar view inside the activity layout
+        Toolbar toolbar = (Toolbar) findViewById(R.id.app_toolbar);
+        // Sets the Toolbar to act as the ActionBar for this Activity window.
+        // Make sure the toolbar exists in the activity and is not null
+        setSupportActionBar(toolbar);
+
+        // Trigger a menu item by setting item clickListener
+        toolbar.setOnMenuItemClickListener(new Toolbar.OnMenuItemClickListener() {
+
+            @Override
+            public boolean onMenuItemClick(MenuItem item) {
+
+                switch (item.getItemId()) {
+                    case R.id.iHome:
+                        startActivity(new Intent(getApplicationContext(), HomeActivity.class));
+                        return true;
+                    case R.id.iShare:
+                        shareIt();
+                        return true;
+                    case R.id.iSet:
+                        setupSetting();
+                        return true;
+                    default:
+                        return onOptionsItemSelected(item);
+                }
+            }
+        });
+    }
+
+
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        createToolBar();
+    }
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+    }
+
+    @Override
+    protected void onPause() {
+        super.onPause();
+    }
+
+
+}
