+++ app/src/main/res/layout/activity_input_interface.xml
@@ -22,7 +22,9 @@
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/app_toolbar">
+        app:layout_constraintTop_toBottomOf="@+id/app_toolbar"
+        app:layout_constraintVertical_bias="0.0"
+        tools:layout_editor_absoluteX="16dp">
 
         <ScrollView
             android:id="@+id/scroll_view"
@@ -32,6 +34,8 @@
             android:background="@color/colorPrimary"
             android:padding="10dp"
             app:layout_constraintBottom_toTopOf="@+id/divInputInterfaceID"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent">
 
             <LinearLayout
